+++ packages/client/base/src/services/api/index.ts
@@ -1,2 +1,1 @@
 export * from "./APIErrorService";
-export * from "./BaseCrossmintService";