+++ packages/client/base/src/services/hosted/index.ts
@@ -1,4 +1,1 @@
-export * from "./crossmintModalService";
-export * from "./crossmintPayButtonService";
-
 export * from "./v3";