+++ packages/client/base/src/utils/i18n/dictionaries/deDE.ts
@@ -1,11 +1,4 @@
 const deDE = {
-    crossmintPayButtonService: {
-        CONNECTING: "Verbindung wird hergestellt...",
-        BUY: "Mit Crossmint kaufen",
-        BUY_WITH_ETH: "Mit ETH kaufen",
-        BUY_WITH_SOL: "Mit SOL kaufen",
-        BUY_WITH_CREDIT_CARD: "Mit Kreditkarte kaufen",
-    },
     hostedCheckoutV3: {
         paymentVariant: {
             pay: "Bezahlen mit {0}",