+++ packages/client/base/src/utils/i18n/dictionaries/jaJp.ts
@@ -1,11 +1,4 @@
 const jaJp = {
-    crossmintPayButtonService: {
-        CONNECTING: "接続中...",
-        BUY: "Crossmintで購入",
-        BUY_WITH_ETH: "ETHで購入",
-        BUY_WITH_SOL: "SOLで購入",
-        BUY_WITH_CREDIT_CARD: "クレジットカードで購入",
-    },
     hostedCheckoutV3: {
         paymentVariant: {
             pay: "{0}で支払う",