+++ packages/client/base/src/utils/i18n/dictionaries/trTR.ts
@@ -1,11 +1,4 @@
 const trTR = {
-    crossmintPayButtonService: {
-        CONNECTING: "Bağlanıyor...",
-        BUY: "Crossmint ile Satın Al",
-        BUY_WITH_ETH: "ETH ile Satın Al",
-        BUY_WITH_SOL: "SOL ile Satın Al",
-        BUY_WITH_CREDIT_CARD: "Kredi Kartı ile Satın Al",
-    },
     hostedCheckoutV3: {
         paymentVariant: {
             pay: "{0} ile öde",