+++ packages/client/base/src/utils/i18n/dictionaries/zhTW.ts
@@ -1,11 +1,4 @@
 const zhTW = {
-    crossmintPayButtonService: {
-        CONNECTING: "連接中...",
-        BUY: "用Crossmint購買",
-        BUY_WITH_ETH: "用ETH購買",
-        BUY_WITH_SOL: "用SOL購買",
-        BUY_WITH_CREDIT_CARD: "用信用卡購買",
-    },
     hostedCheckoutV3: {
         paymentVariant: {
             pay: "使用 {0} 付款",