+++ packages/client/ui/react-ui/src/components/embed/fiat/FiatEmbeddedCheckout.tsx
@@ -1,7 +1,0 @@
-import type { FiatEmbeddedCheckoutProps } from "@crossmint/client-sdk-base";
-
-import FiatEmbeddedCheckoutIFrame from "../../../components/embed/fiat/FiatEmbeddedCheckoutIFrame";
-
-export function CrossmintFiatEmbeddedCheckout(props: FiatEmbeddedCheckoutProps) {
-    return <FiatEmbeddedCheckoutIFrame {...props} />;
-}