+++ packages/client/auth/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @crossmint/client-sdk-auth
 
+## 1.2.28
+
+### Patch Changes
+
+- Updated dependencies [40917df]
+  - @crossmint/client-sdk-base@1.6.0
+  - @crossmint/common-sdk-auth@1.0.50
+
 ## 1.2.27
 
 ### Patch Changes